requirements.txt
alembic==1.4.0
beautifulsoup4==4.8.2
Bootstrap-Flask==1.2.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
google==2.0.3
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jsonify==0.5
Mako==1.1.1
MarkupSafe==1.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
sqlalchemy-datatables==2.0.1
urllib3==1.25.8
Werkzeug==1.0.0
WTForms==2.2.1
psycopg2
configparser
APScheduler==3.6.3
flask-socketio
python-socketio
gevent-websocket
gevent
flask-cors
flask-googlemaps